Use Agent.BuildDirectory instead of Agent.WorkFolder
Agent.WorkFolder is not mounted as docker volume by Azure Pipelines, if
containers are used. This causes the SynopsisDetectTask to fail because
the detect folder cannot be created.

This commit changes the WorkingDirectory to Agent.BuildDirectory. The
detect folder can be created and the detect script can be downloaded
properly even from a container.